Asymetric ice (#719)
* Asymetric ice


* copilot review


---------

+++ stellar-engine/src/stellar_engine/entities/inputs.py
@@ -51,7 +51,7 @@ class ClimateCharge:
     cableTemperature: float
     symmetryType: str
     iceThickness: float
-    frontierSupportNumber: float
+    frontierSupportNumber: int
     iceThicknessBefore: float
     iceThicknessAfter: float
